+++ test/entity_map_test.dart
@@ -3,11 +3,11 @@ import 'package:test/test.dart';
 import 'components.dart';
 
 void main() {
-  test('Map over name', (){
+  test('Map over name', () {
     EntityManager em = EntityManager();
     em.createEntity()..set(Name("Max"))..set(Age(37));
 
-    var nameMap = EntityIndex<Name, String>(em, (name) => name.value );
+    var nameMap = EntityIndex<Name, String>(em, (name) => name.value);
 
     em.createEntity()..set(Name("Alex"))..set(Age(45));
 
@@ -15,17 +15,18 @@ void main() {
     expect(nameMap["Alex"].get<Age>().value, 45);
   });
 
-  test('Multi Map over age', (){
+  test('Multi Map over age', () {
     EntityManager em = EntityManager();
     em.createEntity()..set(Name("Max"))..set(Age(37));
     em.createEntity()..set(Name("Maxim"))..set(Age(45));
 
-    var ageMap = EntityMultiIndex<Age, int>(em, (name) => name.value );
+    var ageMap = EntityMultiIndex<Age, int>(em, (name) => name.value);
 
     em.createEntity()..set(Name("Alex"))..set(Age(37));
 
     expect(ageMap[37].length, 2);
-    expect(ageMap[37].map((e) => e.get<Name>().value), containsAll(["Max", "Alex"]));
+    expect(ageMap[37].map((e) => e.get<Name>().value),
+        containsAll(["Max", "Alex"]));
 
     expect(ageMap[45].length, 1);
     expect(ageMap[45].map((e) => e.get<Name>().value), containsAll(["Maxim"]));
@@ -34,6 +35,7 @@ void main() {
       e.set(Age(45));
     }
     expect(ageMap[45].length, 3);
-    expect(ageMap[45].map((e) => e.get<Name>().value), containsAll(["Maxim", "Max", "Alex"]));
+    expect(ageMap[45].map((e) => e.get<Name>().value),
+        containsAll(["Maxim", "Max", "Alex"]));
   });
 }